+++ cli/tools/fmt.rs
