+++ cli/tools/installer.rs
